python-vision/requirements.txt
opencv-python==4.9.0.80
face-recognition==1.3.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0
pytest==8.2.0
